+++ src/app/api/services/task.service.ts
@@ -188,6 +188,9 @@ export class TaskService extends CachedEntityService<Task> {
         constructorParams: unit,
       },
     ).pipe(
+      map((tasks: Task[]) =>
+        tasks.filter((t) => t.daysSinceSubmission() >= t.unit.feedbackOverflowThresholdDays),
+      ),
       tap((tasks: Task[]) => {
         unit.incorporateTasks(tasks);
       }),
@@ -228,6 +231,7 @@ export class TaskService extends CachedEntityService<Task> {
   public readonly statusSeq = TaskStatus.STATUS_SEQ;
   public readonly helpDescriptions = TaskStatus.HELP_DESCRIPTIONS;
   public readonly statusIcons: Map<TaskStatusEnum, string> = TaskStatus.STATUS_ICONS;
+  public readonly statusMaterialIcons: Map<TaskStatusEnum, string> = TaskStatus.STATUS_MATERIAL_ICONS;
   public readonly rejectFutureStates = TaskStatus.REJECT_FUTURE_STATES;
 
   public statusClass(status: TaskStatusEnum): string {