+++ src/app/doubtfire-angular.module.ts
@@ -52,24 +52,6 @@ import {CodeEditorModule} from '@ngstack/code-editor';
 import {MonacoEditorModule} from 'ngx-monaco-editor-v2';
 
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-import {
-  AboutDoubtfireModal,
-  AboutDoubtfireModalContent,
-} from 'src/app/common/modals/about-doubtfire-modal/about-doubtfire-modal.component';
-import {AboutDoubtfireModalService} from 'src/app/common/modals/about-doubtfire-modal/about-doubtfire-modal.service';
-import {FTaskBadgeComponent} from 'src/app/common/task-badge/task-badge.component';
-import {DoubtfireConstants} from 'src/app/config/constants/doubtfire-constants';
-import {
-  DiscussionComposerDialog,
-  TaskCommentComposerComponent,
-} from 'src/app/tasks/task-comment-composer/task-comment-composer.component';
-import {HttpAuthenticationInterceptor} from './common/services/http-authentication.interceptor';
-import {ProjectTasksListComponent} from './tasks/project-tasks-list/project-tasks-list.component';
-import {
-  D2lUnitDetailsFormComponent,
-  D2lUnitDetailsModal,
-} from './units/states/edit/directives/unit-details-editor/d2l-details-form/d2l-unit-details-form.component';
-
 import {MatButtonToggleModule} from '@angular/material/button-toggle';
 import {
   DateAdapter,
@@ -81,13 +63,27 @@ import {MatDatepickerModule} from '@angular/material/datepicker';
 import {MatProgressSpinnerModule} from '@angular/material/progress-spinner';
 import {MatRadioModule} from '@angular/material/radio';
 import {FlexLayoutModule} from 'ng-flex-layout';
+import {
+  AboutDoubtfireModal,
+  AboutDoubtfireModalContent,
+} from 'src/app/common/modals/about-doubtfire-modal/about-doubtfire-modal.component';
+import {AboutDoubtfireModalService} from 'src/app/common/modals/about-doubtfire-modal/about-doubtfire-modal.service';
+import {FTaskBadgeComponent} from 'src/app/common/task-badge/task-badge.component';
+import {DoubtfireConstants} from 'src/app/config/constants/doubtfire-constants';
+import {
+  DiscussionComposerDialog,
+  TaskCommentComposerComponent,
+} from 'src/app/tasks/task-comment-composer/task-comment-composer.component';
 import {CampusListComponent} from './admin/institution-settings/campuses/campus-list/campus-list.component';
 import {AudioCommentRecorderComponent} from './common/audio-recorder/audio/audio-comment-recorder/audio-comment-recorder';
 import {MicrophoneTesterComponent} from './common/audio-recorder/audio/microphone-tester/microphone-tester.component';
 import {CalendarModalComponent} from './common/modals/calendar-modal/calendar-modal.component';
 import {CommentsModalComponent} from './common/modals/comments-modal/comments-modal.component';
 import {ConfirmationModalComponent} from './common/modals/confirmation-modal/confirmation-modal.component';
+import {DiscussedInClassReasonModalComponent} from './common/modals/discussed-in-class-reason-modal/discussed-in-class-reason-modal.component';
 import {ExtensionModalComponent} from './common/modals/extension-modal/extension-modal.component';
+import {HttpAuthenticationInterceptor} from './common/services/http-authentication.interceptor';
+import {ProjectTasksListComponent} from './tasks/project-tasks-list/project-tasks-list.component';
 import {DiscussionPromptComposerComponent} from './tasks/task-comment-composer/discussion-prompt-composer/discussion-prompt-composer.component';
 import {ExtensionCommentComponent} from './tasks/task-comments-viewer/extension-comment/extension-comment.component';
 import {
@@ -95,6 +91,10 @@ import {
   IntelligentDiscussionPlayerComponent,
 } from './tasks/task-comments-viewer/intelligent-discussion-player/intelligent-discussion-player.component';
 import {IntelligentDiscussionRecorderComponent} from './tasks/task-comments-viewer/intelligent-discussion-player/intelligent-discussion-recorder/intelligent-discussion-recorder.component';
+import {
+  D2lUnitDetailsFormComponent,
+  D2lUnitDetailsModal,
+} from './units/states/edit/directives/unit-details-editor/d2l-details-form/d2l-unit-details-form.component';
 
 import {DateFnsAdapter} from '@angular/material-date-fns-adapter';
 import {DateAdapter as CalendarDateAdapter, CalendarModule} from 'angular-calendar';
@@ -165,6 +165,7 @@ import {UnitStudentsEditorComponent} from './units/states/edit/directives/unit-s
 import {UnitTutorialsListComponent} from './units/states/edit/directives/unit-tutorials-list/unit-tutorials-list.component';
 import {UnitTutorialsManagerComponent} from './units/states/edit/directives/unit-tutorials-manager/unit-tutorials-manager.component';
 import {StudentsListComponent} from './units/states/students-list/students-list.component';
+import {BatchFeedbackWorkflowDialogComponent} from './units/states/tasks/inbox/directives/staff-task-list/batch-feedback-workflow-dialog/batch-feedback-workflow-dialog.component';
 import {StaffTaskListComponent} from './units/states/tasks/inbox/directives/staff-task-list/staff-task-list.component';
 
 import {TaskAssessmentModalComponent} from './common/modals/task-assessment-modal/task-assessment-modal.component';
@@ -185,6 +186,7 @@ import {
 import {FUnitsComponent} from './admin/states/units/units.component';
 import {FUsersComponent} from './admin/states/users/users.component';
 import {TiiActionLogComponent} from './admin/tii-action-log/tii-action-log.component';
+import {AngularRouterHostComponent} from './angular-router-host.component';
 import {FeedbackTemplateService} from './api/services/feedback-template.service';
 import {GroupService} from './api/services/group.service';
 import {LtiService} from './api/services/lti.service';
@@ -230,16 +232,16 @@ import {HttpErrorInterceptor} from './common/services/http-error.interceptor';
 import {SuccessCloseComponent} from './common/success-close/success-close.component';
 import {UnitCodeComponent} from './common/unit-code/unit-code.component';
 import {UserBadgeComponent} from './common/user-badge/user-badge.component';
-import {AngularRouterHostComponent} from './angular-router-host.component';
 import {DoubtfireShellComponent} from './doubtfire-shell.component';
-import {UnauthorisedComponent} from './errors/states/unauthorised/unauthorised.component';
 import {TimeoutComponent} from './errors/states/timeout/timeout.component';
+import {UnauthorisedComponent} from './errors/states/unauthorised/unauthorised.component';
 import {UnavailableCardComponent} from './errors/unavailable-card/unavailable-card.component';
 import {AcceptEulaComponent} from './eula/accept-eula/accept-eula.component';
 import {SplashScreenComponent} from './home/splash-screen/splash-screen.component';
 import {HomeComponent} from './home/states/home/home.component';
 import {LtiDashboardComponent} from './home/states/lti-dashboard/lti-dashboard.component';
 import {LtiUnitLinkComponent} from './home/states/lti-unit-link/lti-unit-link.component';
+import {LegacyRoutePlaceholderComponent} from './legacy-route-placeholder.component';
 import {ProjectProgressDashboardComponent} from './projects/project-progress-dashboard/project-progress-dashboard.component';
 import {StaffNotesViewComponent} from './projects/states/dashboard/directives/task-dashboard/directives/staff-notes-view/staff-notes-view.component';
 import {TaskAssessmentCardComponent} from './projects/states/dashboard/directives/task-dashboard/directives/task-assessment-card/task-assessment-card.component';
@@ -264,6 +266,7 @@ import {TaskFeedbackTemplatesComponent} from './tasks/task-comment-composer/task
 import {ScormCommentComponent} from './tasks/task-comments-viewer/scorm-comment/scorm-comment.component';
 import {ScormExtensionCommentComponent} from './tasks/task-comments-viewer/scorm-extension-comment/scorm-extension-comment.component';
 import {TaskSubmissionHistoryComponent} from './tasks/task-submission-history/task-submission-history.component';
+import {topLevelRoutes, TopLevelUrlHandlingStrategy} from './top-level-routing';
 import {UnitAnalyticsComponent} from './units/states/analytics/unit-analytics-route.component';
 import {TaskDefinitionDatesComponent} from './units/states/edit/directives/unit-tasks-editor/task-definition-editor/task-definition-dates/task-definition-dates.component';
 import {TaskDefinitionEditorComponent} from './units/states/edit/directives/unit-tasks-editor/task-definition-editor/task-definition-editor.component';
@@ -287,8 +290,6 @@ import {FTaskSheetViewComponent} from './units/task-viewer/directives/task-sheet
 import {FUnitTaskListComponent} from './units/task-viewer/directives/unit-task-list/unit-task-list.component';
 import {TaskViewerStateComponent} from './units/task-viewer/task-viewer-state.component';
 import {UnitRootStateComponent} from './units/unit-root-state.component';
-import {LegacyRoutePlaceholderComponent} from './legacy-route-placeholder.component';
-import {topLevelRoutes, TopLevelUrlHandlingStrategy} from './top-level-routing';
 import {ProgressBurndownChartComponent} from './visualisations/progress-burndown-chart/progress-burndown-chart.component';
 import {TaskVisualisationComponent} from './visualisations/task-visualisation/task-visualisation.component';
 import {WelcomeComponent} from './welcome/welcome.component';
@@ -320,11 +321,11 @@ import {FeedbackAppealModalComponent} from './tasks/modals/feedback-appeal-modal
 import {GradeTaskModalComponent} from './tasks/modals/grade-task-modal/grade-task-modal.component';
 import {UploadSubmissionModalComponent} from './tasks/modals/upload-submission-modal/upload-submission-modal.component';
 import {AnalyticsTutorTimesComponent} from './units/states/analytics/directives/analytics-tutor-times.component';
-import {UnitAdminStateComponent} from './units/states/edit/unit-admin-state.component';
 import {UnitDetailsEditorComponent} from './units/states/edit/directives/unit-details-editor/unit-details-editor.component';
 import {UnitStaffEditorComponent} from './units/states/edit/directives/unit-staff-editor/unit-staff-editor.component';
 import {TaskDefinitionDiscussionPromptsComponent} from './units/states/edit/directives/unit-tasks-editor/task-definition-editor/task-definition-discussion-prompts/task-definition-discussion-prompts.component';
 import {OverseerScriptEditorModalComponent} from './units/states/edit/directives/unit-tasks-editor/task-definition-editor/task-definition-overseer/overseer-script-editor-modal/overseer-script-editor-modal.component';
+import {UnitAdminStateComponent} from './units/states/edit/unit-admin-state.component';
 import {DownloadStaffNotesComponent} from './units/states/portfolios/download-staff-notes/download-staff-notes.component';
 import {UploadGradesComponent} from './units/states/portfolios/upload-grades/upload-grades.component';
 import {ConfirmModerationModalComponent} from './units/states/tasks/inbox/directives/moderation/confirm-moderation-modal/confirm-moderation-modal.component';
@@ -341,12 +342,12 @@ import {GroupMemberContributionAssignerComponent} from './groups/group-member-co
 import {GroupMemberListComponent} from './groups/group-member-list/group-member-list.component';
 import {GroupSelectorComponent} from './groups/group-selector/group-selector.component';
 import {GroupSetManagerComponent} from './groups/group-set-manager/group-set-manager.component';
-import {ProjectGroupsComponent} from './projects/states/groups/project-groups/project-groups.component';
 import {ProjectGroupsStateComponent} from './projects/states/groups/project-groups-state.component';
+import {ProjectGroupsComponent} from './projects/states/groups/project-groups/project-groups.component';
 import {PortfolioAddExtraFilesStepComponent} from './projects/states/portfolio/directives/portfolio-add-extra-files-step/portfolio-add-extra-files-step.component';
-import {PortfolioStateComponent} from './projects/states/portfolio/portfolio-state.component';
 import {PortfolioLearningSummaryReportStepComponent} from './projects/states/portfolio/directives/portfolio-learning-summary-report-step/portfolio-learning-summary-report-step.component';
 import {PortfolioWelcomeStepComponent} from './projects/states/portfolio/directives/portfolio-welcome-step/portfolio-welcome-step.component';
+import {PortfolioStateComponent} from './projects/states/portfolio/portfolio-state.component';
 import {TutorialsComponent} from './projects/states/tutorials/tutorials.component';
 import {UnitStudentEnrolmentModalComponent} from './units/modals/unit-student-enrolment-modal/unit-student-enrolment-modal.component';
 import {UnitGroupSetEditorComponent} from './units/states/edit/directives/unit-group-set-editor/unit-group-set-editor.component';
@@ -357,8 +358,8 @@ import {PortfoliosPortfolioViewComponent} from './units/states/portfolios/direct
 import {PortfoliosProjectProgressComponent} from './units/states/portfolios/directives/portfolios-project-progress/portfolios-project-progress.component';
 import {PortfoliosComponent} from './units/states/portfolios/portfolios.component';
 import {RolloverComponent} from './units/states/rollover/rollover.component';
-import {TaskStatusPieChartComponent} from './visualisations/task-status-pie-chart/task-status-pie-chart.component';
 import {UnitTaskInboxStateComponent} from './units/states/tasks/inbox/unit-task-inbox-state.component';
+import {TaskStatusPieChartComponent} from './visualisations/task-status-pie-chart/task-status-pie-chart.component';
 
 // See https://stackoverflow.com/questions/55721254/how-to-change-mat-datepicker-date-format-to-dd-mm-yyyy-in-simplest-way/58189036#58189036
 const MY_DATE_FORMAT = {
@@ -430,6 +431,7 @@ const GANTT_CHART_CONFIG = {
     SpecConModalComponent,
     CalendarModalComponent,
     ConfirmationModalComponent,
+    DiscussedInClassReasonModalComponent,
     InstitutionSettingsComponent,
     ProjectPlanComponent,
     SuccessCloseComponent,
@@ -469,6 +471,7 @@ const GANTT_CHART_CONFIG = {
     SafePipe,
     PdfViewerPanelComponent,
     StaffTaskListComponent,
+    BatchFeedbackWorkflowDialogComponent,
     TaskSimilarityViewComponent,
     TasksForGroupsetPipe,
     OrderByPipe,